Create a user model and login/logout views
Create a url path and a view to display information pages, such as
'about', that takes its template name as a page id. Create a
template for the 'about' page.
Create a model for users that inherits from the AbstractUser model.
Register it in the app's admin.py.
Create views for users to log-in and logout, that inherit from the
LoginView and LogoutView authentications views. Also create a view
that displays a user's profie. Create url paths and templates for
all these views.