Create views for changing user's data and passwords
Create a form that uses the AdvUser model and is used to change
user's data. Then create a view that inherits from the generic
UpdateView view and takes that form. Finally, create a template
for the view.
Create a view that inherits from the PasswordChangeView and is used
to change a user's password. Create a template for the view.