Create models for rubrics objects
Create a model named Rubric and two proxy models that inherit from
it, SuperRubric and Subrubric. Define a Manager class for each of
these proxy models and change their ordering behavior.
In order to place a variable, containing a list of all Subrubrics,
inside of each template's context, create a custom context-
processor named bboard_context_processor, and add it to the
TEMPLATES setting inside the settings.py file.
Change the base.html temlate and make it display rubrics dynamically.